+++ scico/test/linop/test_stack.py
@@ -5,17 +5,11 @@ import jax
 import pytest
 
 import scico.numpy as snp
-from scico.linop import (
-    BlockDiagonalLinearOperator,
-    Convolve,
-    Identity,
-    LinearOperatorStack,
-    Sum,
-)
+from scico.linop import Convolve, DiagonalStack, Identity, Sum, VerticalStack
 from scico.test.linop.test_linop import adjoint_test
 
 
-class TestLinearOperatorStack:
+class TestVerticalStack:
     def setup_method(self, method):
         self.key = jax.random.PRNGKey(12345)
 
@@ -24,18 +18,18 @@ class TestLinearOperatorStack:
         # requires a list of LinearOperators
         I = Identity((42,))
         with pytest.raises(ValueError):
-            H = LinearOperatorStack(I, jit=jit)
+            H = VerticalStack(I, jit=jit)
 
         # checks input sizes
         A = Identity((3, 2))
         B = Identity((7, 2))
         with pytest.raises(ValueError):
-            H = LinearOperatorStack([A, B], jit=jit)
+            H = VerticalStack([A, B], jit=jit)
 
         # in general, returns a BlockArray
         A = Convolve(jax.device_put(np.ones((3, 3))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], jit=jit)
+        H = VerticalStack([A, B], jit=jit)
         x = np.ones((7, 11))
         y = H @ x
         assert y.shape == ((9, 13), (8, 12))
@@ -47,7 +41,7 @@ class TestLinearOperatorStack:
         # by default, collapse to DeviceArray when possible
         A = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], jit=jit)
+        H = VerticalStack([A, B], jit=jit)
         x = np.ones((7, 11))
         y = H @ x
         assert y.shape == (2, 8, 12)
@@ -59,7 +53,7 @@ class TestLinearOperatorStack:
         # let user turn off collapsing
         A = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], collapse=False, jit=jit)
+        H = VerticalStack([A, B], collapse=False, jit=jit)
         x = np.ones((7, 11))
         y = H @ x
         assert y.shape == ((8, 12), (8, 12))
@@ -70,13 +64,13 @@ class TestLinearOperatorStack:
         # general case
         A = Convolve(jax.device_put(np.ones((3, 3))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], collapse=collapse, jit=jit)
+        H = VerticalStack([A, B], collapse=collapse, jit=jit)
         adjoint_test(H, self.key)
 
         # collapsable case
         A = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], collapse=collapse, jit=jit)
+        H = VerticalStack([A, B], collapse=collapse, jit=jit)
         adjoint_test(H, self.key)
 
     @pytest.mark.parametrize("collapse", [False, True])
@@ -85,11 +79,11 @@ class TestLinearOperatorStack:
         # adding
         A = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
         B = Convolve(jax.device_put(np.ones((2, 2))), (7, 11))
-        H = LinearOperatorStack([A, B], collapse=collapse, jit=jit)
+        H = VerticalStack([A, B], collapse=collapse, jit=jit)
 
         A = Convolve(jax.device_put(np.random.rand(2, 2)), (7, 11))
         B = Convolve(jax.device_put(np.random.rand(2, 2)), (7, 11))
-        G = LinearOperatorStack([A, B], collapse=collapse, jit=jit)
+        G = VerticalStack([A, B], collapse=collapse, jit=jit)
 
         x = np.ones((7, 11))
         S = H + G
@@ -101,11 +95,11 @@ class TestLinearOperatorStack:
         np.testing.assert_allclose((S @ x)[1], (H @ x + G @ x)[1])
 
         # result of adding two conformable stacks should be a stack
-        assert isinstance(S, LinearOperatorStack)
-        assert isinstance(H - G, LinearOperatorStack)
+        assert isinstance(S, VerticalStack)
+        assert isinstance(H - G, VerticalStack)
 
         # scalar multiplication
-        assert isinstance(1.0 * H, LinearOperatorStack)
+        assert isinstance(1.0 * H, VerticalStack)
 
         # op scaling
         scalars = [2.0, 3.0]
@@ -124,7 +118,7 @@ class TestBlockDiagonalLinearOperator:
         A1 = Identity(S1)
         A2 = 2 * Identity(S2)
         A3 = Sum(S3)
-        H = BlockDiagonalLinearOperator((A1, A2, A3))
+        H = DiagonalStack((A1, A2, A3))
 
         x = snp.ones((S1, S2, S3))
         y = H @ x
@@ -139,7 +133,7 @@ class TestBlockDiagonalLinearOperator:
         A1 = Identity(S1)
         A2 = 2 * Identity(S2)
         A3 = Sum(S3)
-        H = BlockDiagonalLinearOperator((A1, A2, A3))
+        H = DiagonalStack((A1, A2, A3))
 
         y = snp.ones((S1, S2, ()), dtype=snp.float32)
         x = H.T @ y
@@ -158,10 +152,10 @@ class TestBlockDiagonalLinearOperator:
         A1 = Identity(S)
         A2 = Sum(S)
 
-        H = BlockDiagonalLinearOperator((A1, A2))
+        H = DiagonalStack((A1, A2))
         assert H.input_shape == (2, *S)
 
-        H = BlockDiagonalLinearOperator((A1, A2), allow_input_collapse=False)
+        H = DiagonalStack((A1, A2), allow_input_collapse=False)
         assert H.input_shape == (S, S)
 
     def test_output_collapse(self):
@@ -170,8 +164,8 @@ class TestBlockDiagonalLinearOperator:
         A1 = Identity(S1)
         A2 = Sum(S2, axis=0)
 
-        H = BlockDiagonalLinearOperator((A1, A2))
+        H = DiagonalStack((A1, A2))
         assert H.output_shape == (2, *S1)
 
-        H = BlockDiagonalLinearOperator((A1, A2), allow_output_collapse=False)
+        H = DiagonalStack((A1, A2), allow_output_collapse=False)
         assert H.output_shape == (S1, S1)